Sample Size Determination for Stratified Sampling

The function stratasize determinates the total size of stratified samples depending on type of allocation and determinated by specified precision.

stratasize(e, Nh, Sh, level = 0.95, type = 'prop')

Arguments

e
positive number specifying sampling precision.

Nh
vector of population sizes in each stratum.

Sh
vector of standard deviation in each stratum.

level
coverage probability for confidence intervals. Default is level=0.95.

type
type of allocation. Default is type='prop' for proportional, alternative is type='opt' for optimal.

Value

The function stratasize returns a value, which is a list consisting of the components

call

is a list of call components: e specified precision, Nh population sizes of every stratum, Sh standard diviation of every stratum, method type of allocation, level coverage probability for confidence intervals.

n

determinated total sample size.

References

Kauermann, Goeran/Kuechenhoff, Helmut (2011): Stichproben. Methoden und praktische Umsetzung mit R. Springer.

See also

Examples

#random proportional stratified sample
stratasize(e=0.1, Nh=c(100000,300000,600000), Sh=c(1,2,3))

#> 
#> stratamean object: Stratified sample size determination
#> 
#> type of sample: prop
#> 
#> total sample size determinated: 2568

#random optimal stratified sample
stratasize(e=0.1, Nh=c(100000,300000,600000), Sh=c(1,2,3), type="opt")

#> 
#> stratamean object: Stratified sample size determination
#> 
#> type of sample: opt
#> 
#> total sample size determinated: 2395