Chinese wage data

A data frame with hypothetical Chinese wages differenciated by region and industrial sector.

data(wage)

Format

A data frame with 231 observations on the following 3 variables.

Region

factor, Chinese regions with 14 levels.

Sector

factor, industrial sector with 30 levels.

Wage

a numeric vector, average wage in the region and sector measured in Chinese yuan.

Details

The dataset is hypothetical. Its structure imitates the data in the Chinese Statistical Yearbook. The values are simulated corresponding to the distribution of the real data which are not publicly accessible.

References

Kauermann, Goeran/Kuechenhoff, Helmut (2010): Stichproben. Methoden und praktische Umsetzung mit R. Springer.

Examples

# Chinese wage data
data(wage)
summary(wage)

#>        Region                 Sector         Wage      
#>  Henan    : 19   BusinessServices: 11   Min.   : 4698  
#>  Liaoning : 19   Husbandry       : 11   1st Qu.:13918  
#>  Chongqing: 18   Construction    : 10   Median :16695  
#>  Beijing  : 17   Financial       : 10   Mean   :16760  
#>  Hubei    : 17   Health          : 10   3rd Qu.:19533  
#>  Jiangsu  : 17   Research        : 10   Max.   :29530  
#>  (Other)  :124   (Other)         :169                  

submean(wage$Wage,PSU=wage$Region, N=990, M=33, Nl=rep(30,14))

#> 
#> submean object: Sub-sample mean estimate
#> With finite population correction.
#> Using method: simple
#> 
#> Mean estimate: 16753.19
#> Standard error: 475.3202
#> 95% confidence interval: [15821.58,17684.8]
#>