Sample Size Calculation for Stratified Sampling

The function stratasamp calculates the sample size for each stratum depending on type of allocation.

stratasamp(n, Nh, Sh = NULL, Ch = NULL, type = 'prop')

Arguments

n
positive integer specifying sampling size.

Nh
vector of population sizes of each stratum.

Sh
vector of standard deviation in each stratum.

Ch
vector of cost for a sample in each stratum.

type
type of allocation. Default is type='prop' for proportional, alternatives are type='opt' for optimal and type='costopt' for cost-optimal.

Value

The function stratasamp returns a matrix, which lists the strata and the sizes of observation depending on type of allocation.

References

Kauermann, Goeran/Kuechenhoff, Helmut (2010): Stichproben. Methoden und praktische Umsetzung mit R. Springer.

See also

Examples

#random proportional stratified sample
stratasamp(n=500, Nh=c(5234,2586,649,157))

#>                     
#> Stratum   1   2  3 4
#> Size    303 150 38 9

stratasamp(n=500, Nh=c(5234,2586,649,157), Sh=c(251,1165,8035,24725), type='opt')

#>                       
#> Stratum  1   2   3   4
#> Size    49 112 194 145